+++ test/raxx_test.exs
@@ -10,6 +10,24 @@ defmodule RaxxTest do
     end
   end
 
+  test "header values cannot contain control feed charachter" do
+    assert_raise RuntimeError,
+                 "Header values must not contain control feed (\\r) or newline (\\n)",
+                 fn ->
+                   Raxx.response(:ok)
+                   |> Raxx.set_header("foo", "Bar\r")
+                 end
+  end
+
+  test "header values cannot contain newline charachter" do
+    assert_raise RuntimeError,
+                 "Header values must not contain control feed (\\r) or newline (\\n)",
+                 fn ->
+                   Raxx.response(:ok)
+                   |> Raxx.set_header("foo", "Bar\n")
+                 end
+  end
+
   test "cannot set a header twice" do
     assert_raise RuntimeError, "Headers should not be duplicated", fn ->
       Raxx.response(:ok)